pyproject.toml
[project]
name = "quote-aqi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kainan Zhu", email = "dlangu0393@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "icecream>=2.1.8",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
]

[project.scripts]
quote-aqi = "quote_aqi:main"

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"
